requirements.txt
# --- choose ONE of these two lines ---
# strands-agents[openai]>=1.7.1      # If using OpenAI models via OpenAI API
strands-agents            # If using Ollama/Bedrock/etc. (no OpenAI extras)

# Optional/handy tool pack
strands-agents-tools

# Env & typing niceties (optional but recommended)
python-dotenv
pydantic
